Move cancellation CLI commands to their forever home
